refactor: 출력 기능을 OutputView 파일로 분리
